Basic Setup Improvements (#18)
- Add to .gitignore to avoid devs committing user-related files
- Update README to improve clarity and reference new setup-mac script
- convert cpu/memory options to directly use terraform `-var` option
- move lambda-related files to their own folder
- add type constraints to terraform variables
- replace make_lambda_zip script with terraform archive_file resource
- add environment setup scripts for mac, linux, and windows
